move types into file
